+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Add ``icon-right`` prop to autocomplete
 * Add ``vertical`` and ``rounded`` props to steps (thanks @service-paradis)
 * Add ``step`` prop to step item (thanks @service-paradis)
+* Add ``card-layout`` prop to table
 
 ### Fixes